Fix potential overflow with digital

+++ src/main/java/org/example/concurrent/BlockingBag.java
@@ -1,8 +1,8 @@
 package org.example.concurrent;
 
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public interface BlockingBag {
-  void add(int element);
-  IntStream takes(int countElements) throws InterruptedException;
+  void add(long element);
+  LongStream takes(int countElements) throws InterruptedException;
 }
